face-service/requirements.txt
pymongo>=4.0.0
python-dotenv>=1.0.0
boto3>=1.28.0
